Changed naming of categories, fixed pivots
(1) The old codes were calculated on the basis of gauge-invariants while it should have been on the basis of full invariants.
This has now been fixed and new codes were assigned.
This means the order of the categories has been changed.
(2) The pivotal data was the inverse of what it should be. This is updated now. For most categories this has no effects.
Note: the left quantum dimensions were correct.