Changed ordering fusion cats, fixed pivots
There was a mistake in the algorithm for determining the codes for the fusion cats: it was dependent on fusion ring automorphisms. This has been fixed and the codes have been updated.
The pivots of the categories were also inverted. This has been fixed as well. Since all quantum dimensions were still correct, this has no effect on the naming or properties of the categories. Most pivots were +-1 so only very few pivotal structures needed to be corrected